Seção 17- Palestrantes e redes sociais Frontend

+++ Back/src/ProEventos.API/helpers/Util.cs
@@ -32,9 +32,12 @@ namespace ProEventos.API.helpers
         }
         public void DeleteImage(string imageName, string destino)
         {
-            var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @$"Resources/{destino}", imageName);
-            if (File.Exists(imagePath))
-                File.Delete(imagePath);
+            if (!string.IsNullOrEmpty(imageName))
+            {
+                var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @$"Resources/{destino}", imageName);
+                if (File.Exists(imagePath))
+                    File.Delete(imagePath);
+            }
         }
 
     }
